Fixes a bug in SafraFT.
If two tokens are received for the same round it clears the crashed set of the second token which is then forwarded empty.
This is a problem because some nodes will never learn that another has crashed.
Now only nodes that have been reported crashed by a token from the last round are removed.